+++ app/src/main/java/org/tensorflow/lite/examples/classification/Presentation/MaskDetectionActivity/ClassifierActivity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.tensorflow.lite.examples.classification.Presentation;
+package org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity;
 
 import android.graphics.Bitmap;
 import android.graphics.Bitmap.Config;
@@ -27,11 +27,11 @@ import android.util.TypedValue;
 import android.widget.Toast;
 import java.io.IOException;
 import java.util.List;
-import org.tensorflow.lite.examples.classification.Presentation.env.BorderedText;
-import org.tensorflow.lite.examples.classification.Presentation.env.Logger;
-import org.tensorflow.lite.examples.classification.Presentation.tflite.Classifier;
-import org.tensorflow.lite.examples.classification.Presentation.tflite.Classifier.Device;
-import org.tensorflow.lite.examples.classification.Presentation.tflite.Classifier.Model;
+import org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.env.BorderedText;
+import org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.env.Logger;
+import org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.tflite.Classifier;
+import org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.tflite.Classifier.Device;
+import org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.tflite.Classifier.Model;
 import org.tensorflow.lite.examples.classification.R;
 
 public class ClassifierActivity extends CameraActivity implements OnImageAvailableListener {
