+++ app/src/main/java/org/tensorflow/lite/examples/classification/Presentation/MaskDetectionActivity/customview/AutoFitTextureView.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.tensorflow.lite.examples.classification.Presentation.customview;
+package org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.customview;
 
 import android.content.Context;
 import android.util.AttributeSet;
