+++ app/src/main/java/org/tensorflow/lite/examples/classification/Presentation/MaskDetectionActivity/customview/OverlayView.java
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-package org.tensorflow.lite.examples.classification.Presentation.customview;
+package org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.customview;
 
 import android.content.Context;
 import android.graphics.Canvas;
