+++ app/src/main/java/org/tensorflow/lite/examples/classification/Presentation/MaskDetectionActivity/customview/ResultsView.java
@@ -13,10 +13,10 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-package org.tensorflow.lite.examples.classification.Presentation.customview;
+package org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.customview;
 
 import java.util.List;
-import org.tensorflow.lite.examples.classification.Presentation.tflite.Classifier.Recognition;
+import org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.tflite.Classifier.Recognition;
 
 public interface ResultsView {
   public void setResults(final List<Recognition> results);
