+++ app/src/main/java/org/tensorflow/lite/examples/classification/Presentation/MaskDetectionActivity/tflite/ClassifierFloatMobileNet.java
@@ -1,4 +1,4 @@
-/* Copyright 2017 The TensorFlow Authors. All Rights Reserved.
+/* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-package org.tensorflow.lite.examples.classification.Presentation.tflite;
+package org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.tflite;
 
 import android.app.Activity;
 import java.io.IOException;
@@ -21,28 +21,28 @@ import java.io.IOException;
 import org.tensorflow.lite.support.common.TensorOperator;
 import org.tensorflow.lite.support.common.ops.NormalizeOp;
 
-/** This TensorFlow Lite classifier works with the quantized MobileNet model. */
-public class ClassifierQuantizedMobileNet extends Classifier {
+/** This TensorFlowLite classifier works with the float MobileNet model. */
+public class  ClassifierFloatMobileNet extends Classifier {
+
+  /** Float MobileNet requires additional normalization of the used input. */
+  private static final float IMAGE_MEAN = 127.5f;
+
+  private static final float IMAGE_STD = 127.5f;
 
   /**
-   * The quantized model does not require normalization, thus set mean as 0.0f, and std as 1.0f to
-   * bypass the normalization.
+   * Float model does not need dequantization in the post-processing. Setting mean and std as 0.0f
+   * and 1.0f, repectively, to bypass the normalization.
    */
-  private static final float IMAGE_MEAN = 0.0f;
-
-  private static final float IMAGE_STD = 1.0f;
-
-  /** Quantized MobileNet requires additional dequantization to the output probability. */
   private static final float PROBABILITY_MEAN = 0.0f;
 
-  private static final float PROBABILITY_STD = 255.0f;
+  private static final float PROBABILITY_STD = 1.0f;
 
   /**
-   * Initializes a {@code ClassifierQuantizedMobileNet}.
+   * Initializes a {@code ClassifierFloatMobileNet}.
    *
    * @param activity
    */
-  public ClassifierQuantizedMobileNet(Activity activity, Device device, int numThreads)
+  public ClassifierFloatMobileNet(Activity activity, Device device, int numThreads)
       throws IOException {
     super(activity, device, numThreads);
   }
@@ -52,7 +52,7 @@ public class ClassifierQuantizedMobileNet extends Classifier {
     // you can download this file from
     // see build.gradle for where to obtain this file. It should be auto
     // downloaded into assets.
-    return "model.tflite";
+    return "model_unquant.tflite";
   }
 
   @Override
