+++ app/src/main/res/layout/activity_camera.xml
@@ -28,7 +28,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        tools:context="org.tensorflow.lite.examples.classification.Presentation.CameraActivity" />
+        tools:context="org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.CameraActivity" />
 
     <LinearLayout
         android:layout_width="match_parent"
