+++ app/src/main/res/layout/camera_connection_fragment.xml
@@ -17,7 +17,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <org.tensorflow.lite.examples.classification.Presentation.customview.AutoFitTextureView
+    <org.tensorflow.lite.examples.classification.Presentation.MaskDetectionActivity.customview.AutoFitTextureView
         android:id="@+id/texture"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
